+++ CHANGELOG.rst
@@ -28,6 +28,11 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Unreleased
 ------------------------------------------------------------------------------
 
+* |Feature| Add support for typed CSVs (option ``dtypes_path`` of
+  :class:`FileDataSource <mllaunchpad.datasources.FileDataSource>` and
+  :class:`FileDataSink <mllaunchpad.datasources.FileDataSink>`),
+  `issue #127 <https://github.com/schuderer/mllaunchpad/issues/127>`_,
+  by `Elisa Partodikromo <https://github.com/planeetjupyter>`_.
 * |Fixed| Keep generated RAML files free of command line messages,
   `issue #126 <https://github.com/schuderer/mllaunchpad/issues/126>`_,
   by `Andreas Schuderer <https://github.com/schuderer>`_.